requirements.txt
adrf==0.1.9
asgiref==3.8.1
async-property==0.2.2
attrs==25.3.0
Django==5.1.8
django-environ==0.12.0
django-filter==25.1
django-tenants==3.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-writable-nested==0.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
psycopg==3.2.6
psycopg-binary==3.2.6
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
